requirements.txt
google-cloud-bigquery==3.13.0
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.0
zohocrmsdk2-1==3.0.0
schedule==1.2.0